fix: 修复3个编译错误
1. RouteActivity.kt - Coil MemoryCache maxSizeBytes Long 类型适配
2. DataSourceModule.kt - 恢复 TimeUnit 导入（Ktor engine 需要）
3. Mermaid.kt - 添加 LocalDensity 缺失导入
4. build.gradle.kts - 移除废弃的 -Xjvm-default 参数

+++ app/src/main/java/me/rerere/rikkahub/RouteActivity.kt
@@ -175,15 +175,11 @@ class RouteActivity : ComponentActivity() {
         setContent {
             RikkahubTheme {
                 setSingletonImageLoaderFactory { context ->
-                    val memoryCacheSize = (Runtime.getRuntime().maxMemory() / 4).toInt() // 25% of heap = ~128MB
+                    val memoryCacheSize = Runtime.getRuntime().maxMemory() / 4 // 25% of heap = ~128MB
                     ImageLoader.Builder(context)
                         .crossfade(true)
                         .memoryCachePolicy(CachePolicy.ENABLED)
-                        .memoryCache {
-                            MemoryCache.Builder()
-                                .maxSizeBytes(memoryCacheSize)
-                                .build()
-                        }
+                        .memoryCache { MemoryCache.Builder().maxSizeBytes(memoryCacheSize).build() }
                         .components {
                             add(
                                 OkHttpNetworkFetcherFactory(
